rewrote more parts of the InteractionHandler class and created LockManager.java and GateManager.java
